Created gaeTag; Improved gaePlayer
gaeTag will allow us to add more control to the generation of files in the future by adding tags to the loop pieces within the Layer Folders.

gaePlayer recieves a new seed based generation, so we can guaranteed generate the same arrangment if need be. Also changed it so that the cover art is resized to 500 pixels in height.